Update files for bibl reconciliation
Many fixes to bibl ids, modifications to some ids, to agree with
titles.xml now; should clear most of the issues for #25 (and also clears
up a number of issues from #22). There are still a few remaining things
in titles.xml that aren't cited, but without tracking them down I can't
be sure which inscriptions they're there for.